Collapse extensions into Score.swift to avoid build order bug
